requirements.txt
google-genai
python-docx
python-dotenv
pydantic
fastapi
uvicorn
python-multipart
pytest
httpx
pytest-mock